fix(team): correct the SCJoinedTeam wire layout and widen member ids
Ported from fd285fe4, fb8ca8cc and ff1339ae. None of these files, nor
the opcode, were touched on 10.0.2_r575:

- SCJoinedTeam is 0x10B on 10.0.2.13, not 0xd6, and its body layout was
  wrong - parties and raids never formed
- member ids are 8 bytes in SCRefreshTeamMember,
  SCTeamMemberDisconnected and SCTeamMemberRoleChanged
- party members keep their map marker once they leave visual range

The Solo* GM commands from fd285fe4 are not in this commit; they need
the ExpeditionManager and FamilyManager changes that go with the guild
work.

+++ AAEmu.Game/Core/Packets/G2C/SCRefreshTeamMemberPacket.cs
@@ -3,14 +3,17 @@ using AAEmu.Game.Core.Network.Game;
 
 namespace AAEmu.Game.Core.Packets.G2C;
 
-public class SCRefreshTeamMemberPacket(uint teamId, uint memberId, uint objId)
+/// <summary>
+/// Client layout (VA 0x39C79880): tid u32, type u64, bc. The member id is EIGHT bytes.
+/// </summary>
+public class SCRefreshTeamMemberPacket(uint teamId, ulong memberId, uint objId)
     : GamePacket(SCOffsets.SCRefreshTeamMemberPacket, 1)
 {
     public override PacketStream Write(PacketStream stream)
     {
-        stream.Write(teamId);
-        stream.Write(memberId);
-        stream.WriteBc(objId);
+        stream.Write(teamId);       // u32 tid
+        stream.Write(memberId);     // u64 type
+        stream.WriteBc(objId);      // bc
         return stream;
     }
 }
